Update to Django 3.1 and django-mapit 3.0
Since last running these scripts, I've upgraded Ubuntu to 20.04 (Focal)
which has GDAL 3. This version of GDAL reverses lon,lat to lat,lon for
WGS84 which breaks everything - see:

  https://code.djangoproject.com/ticket/31695#no1

The solution is to update to Django 3.1, which works with GDAL 3.
django-mapit must also be updated to a version that works with Django
3.1.