Fix for making postcode polygons for cross-region postcodes
The code is now calculating the Voronoi diagram per EUR region, but the
script that made the final postcode polygons was cropping them to the
union of EUR regions that points with that postcode lie in. This meant
that for a cross-border postcode (e.g. on the North East / Scotland
border) you could have a Voronoi region for a point just on the English
side which extended to the points-at-infinity way to the north of
Scotland, but they would be cropped by the union of the North East and
Scotland, so most of Scotland would be included in that postcode's
multi-polygon.

To fix this, this commit changes the behaviour when forming postcode
polygons to crop by the source region of the points before unioning the
Voronoi regions arising from each region. Hopefully this will fix the
problem.