Allow restricting data generation only by area
Previously this file had some code to restrict the scope by an arbitrary
prefix, but that actually produced very odd results. Instead this commit
changes it so that you can only generate a slice of the data for a
single area, if you're not generating everything. This has the
advantages that it (a) actually works and (b) still saves a lot of time
if you need to correct just a few postcodes.